+++ src/codec/adau1361lower.cpp
@@ -511,7 +511,7 @@ void ::rpp_driver::Adau1361Lower::InitializeRegisters() {
 }
 
 // Set the converter clock.
-void ::rpp_driver::Adau1361Lower::ConfigureSRC(unsigned int fs) {
+void ::rpp_driver::Adau1361Lower::ConfigureSrc(unsigned int fs) {
   assert((fs == 24000 || fs == 32000 || fs == 48000 || fs == 96000 ||
           fs == 22050 || fs == 44100 || fs == 88200) &&
          "Bad Fs");