+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 History of the project development
 
 ## [Unreleased] yyyy-mm-zz
+The issue #16 changes its public member function. But it is referred internally from ::rpp_driver::Adau1361 class. Not called by user program. Then, the major version number is not needed to change. 
+
 ### Added
 ### Changed
 - [Issue #10](https://github.com/suikan4github/rpp_driver/issues/10) Cover APIs of the Raspberry Pi Pico SDK.
+- [Issue #16](https://github.com/suikan4github/rpp_driver/issues/16)  Change the member function name from ConfigureSRC() to ConfigureSrc().
 ### Deprecated
 ### Removed
 ### Fixed